Migrate alias analysis test cases
